\name{Artificial}
\docType{data}
\alias{Artificial}
\title{Two sets of 3-part compositions}
\description{
   The list object \code{Artificial} contains two data frames of 3-part
   compositions. The data refer to the example in Section 3.1 of
   Graffelman et al. (2017) 
}
\usage{data(Artificial)}
\format{A list containing two data frames containing 100 observations.}
\source{Laird, N. M. and Lange, C. Table 7.11, p. 124}
\references{Graffelman, J., Pawlowsky-Glahn, V., Egozcue, J.J. and
  Buccianti, A. (2017) Compositional Canonical Correlation Analysis. 
}
\keyword{datasets}
     