\name{Tubb}
\alias{Tubb}
\docType{data}
\title{Romano-British pottery oxides}
\description{
A dataframe with the major oxide composition of pottery found at Romano-British kiln sites in Wales,
Gloucester and the New Forest as determined by atomic absorption.
}
\usage{data("Tubb")}
\format{
  A data frame with 48 observations on the following 11 variables.
  \describe{
    \item{\code{Sample}}{Sample identifier}
    \item{\code{Al2O3}}{Aluminium oxide}
    \item{\code{Fe2O3}}{Iron (III) oxide}
    \item{\code{MgO}}{Magnesium oxide}
    \item{\code{CaO}}{Calcium oxide}
    \item{\code{Na2O}}{Sodium oxide}
    \item{\code{K2O}}{Potassium oxide}
    \item{\code{TiO2}}{Titaniium dioxide}
    \item{\code{MnO}}{Manganese oxide}
    \item{\code{BaO}}{Barium oxide}
    \item{\code{site}}{Geographical region of the sample. G=Gloucester, NF=New Forest, W=Wales.}
  }
}
\references{
Tubb, A., Parker, A.J. and Nickless, G. (1980) The analysis of Romano-British pottery by atomic absorption spectrophotometry. 
Archaeometry 22(2) pp. 153--171.
}
\examples{
data(Tubb)
}
\keyword{datasets}
